move TestParameters to utils
