+++ tests/integration/base_tests.py
@@ -2,7 +2,6 @@
 Base classes for integration tests
 """
 
-import dataclasses
 import logging
 import os
 import re
@@ -28,25 +27,6 @@ from python2verilog.simulation import iverilog
 from python2verilog.simulation.display import strip_ready, strip_valid
 
 from .utils import make_tuple
-
-
-@dataclasses.dataclass
-class TestParameters:
-    # Function to-be-tested
-    func: FunctionType
-
-    # List of argument combinations to-be-tested
-    # Each element is a separate test case
-    # Each element is unpacked as the arguments for that test case
-    args_list: list[Union[tuple[int, ...], int]]
-
-    # All functions called by `func`
-    helpers: list[FunctionType]
-
-    # Optimization levels to-be-tested
-    opti_levels: list[int] = dataclasses.field(default_factory=lambda: [0, 1, 2, 4, 8])
-
-    # This results in `len(args_list) * len(opti_levels)` test cases
 
 
 class BaseTestWrapper:
