+++ tests/integration/utils.py
@@ -1,3 +1,8 @@
+"""
+Utilities for integration tests
+"""
+
+import dataclasses
 from types import FunctionType
 from typing import Union
 
@@ -22,3 +27,22 @@ def name_func(testcase_func: FunctionType, _: int, param: dict) -> str:
     else:
         # If given list of funcs, use first for naming
         return f"{testcase_func.__name__}::{param.args[0][0].__name__}"
+
+
+@dataclasses.dataclass
+class TestParameters:
+    # Function to-be-tested
+    func: FunctionType
+
+    # List of argument combinations to-be-tested
+    # Each element is a separate test case
+    # Each element is unpacked as the arguments for that test case
+    args_list: list[Union[tuple[int, ...], int]]
+
+    # All functions called by `func`
+    helpers: list[FunctionType]
+
+    # Optimization levels to-be-tested
+    opti_levels: list[int] = dataclasses.field(default_factory=lambda: [0, 1, 2, 4, 8])
+
+    # This results in `len(args_list) * len(opti_levels)` test cases
